backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-docx>=1.1.0
pymupdf>=1.23.0
spacy>=3.7.0
sentence-transformers>=2.2.0
pandas>=2.1.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
loguru>=0.7.0
xgboost>=2.0.0
joblib>=1.3.0
faker>=20.0.0
httpx>=0.25.0
